requirements.txt
flask
edge-tts
google-generativeai
openai
python-dotenv